Create dimension table for usage category and use it for queries
Summary:
This creates a dimension table for usage that maps a BI to all the
categories they are members of. This table is then referenced in the
dataset for usage to execute queries.

This changeset also includes handling of `NULL` filter i.e
```
{
  "field_name": "sub_account_sid",
  "equality": "equal",
  "field_value": null
}
```

For cube tables where `sub_account_sid`, this translates the filter to
```sql
sub_account_sid = 'IS_NULL'
```

RC-834