Signal scan is finished with a Message instead of AtomicInteger
Summary:
There appears to be a race condition in the current implementation
that can cause the Enumerable to return false even when there are
objects on the queue. It seems this happens because we were using two
concurrency primitives, a Queue and a AtomicInteger. This makes it
just use the Queue and passes Messages into the queue. Each Message is
one of three types:

1. ROW
2. CLOSE
3. ERROR

ROW indicates it has a set of values to return to the top level
enumerable, CLOSE indicates the scanner is complete. ERROR includes an
exception. Sadly, when this happens we will create a deep stack trace
that will be harder to parse. I hope we will be ok.

RC-906