Move to Unbounded Queues
Summary:
Using Unbounded Queues here which is not awesome.
Attempted to come up with a few ways in which the producer could
block when the queue was full didn't really like anything I had
written. The point of using these queues is to avoid blocking
all together. Further, because the Queue is used as the message
passing for producer errors, it seemed like we needed another signal
to send to the Enumerable. When this gets a refactor for the Default
Sort order, lets address the "unbounded-ness" of this and add a
failure signal to the consumer so it doesn't rely on a single Message
Passing Channel.