Increase bindable cache hit rate
Summary:
So after exploring a few things within Calcite I finally figured out
what `stash` API does and it is kindof exactly what we needed. So when
using `stash` on non-primatives, the object passed into the method
call is inserted into `DataContext`s map. So the code looks like this:

```java
/*   1 */ public org.apache.calcite.linq4j.Enumerable bind(final org.apache.calcite.DataContext root) {
/*   2 */   final java.util.List v1stashed = (java.util.List) root.get("v1stashed");
/*   3 */   final java.util.List v2stashed = (java.util.List) root.get("v2stashed");
/*   4 */   return ((com.twilio.raas.sql.CalciteKuduTable.KuduQueryable) org.apache.calcite.schema.Schemas.queryable(root, root.getRootSchema().getSubSchema("kudu"), java.lang.Object[].class, "JoinedCommunications-DailyUsage-Aggregation")).query(v1stashed, v2stashed, 2L, -1L, false);
/*   5 */ }
/*   6 */
/*   7 */
/*   8 */ public Class getElementType() {
/*   9 */   return java.lang.Object[].class;
/*  10 */ }
/*  11 */
/*  12 */
```

This is perfect, it means for the same query with the same predicates
generates the same java code, but the `DataContext` will hold the
predicates that are different.

RC-1086