Allow descending datetime translation for audit events
Fix jar to run for audit event mock data

Update Audit events mock-data to include dimension table

Force flush dimension table for audit events

Allow reverse dateTime translation

- This commit allows us to store records in most recent on dateTime format
- As this causes the datetime to not be stored in sql.Timestamp, Calcite's physical layer handles the translation to convert datetimes back into sql.Timestamp

Changes as per review and other fixes for descending sort for datetime

- Fix Predicates so that logical layer always thinks of datetime as UNIXTIME_MICROS and not LONG(in case of desc stored dates)
- Change name from reverseSortedDateTimeField to descendingSortedDateTimeField
- Add in IT to test all possible cases for descending sort on datetime

Make audit_events mock data include sub_account_sid

Subaccount filter is same as account sid treat as `IS_NULL`

Summary:
This is a regression in the data that messaging team built into their
interface. This addresses that regression by acknowledging that a
filter that doesn't make strict sql sense -- you can't have the record
be created by a sub account and an parent account that have the same
string value -- and turn it into something that makes sense.

[job-creator] prepare release release-2.11.6

[job-creator] prepare for next development iteration

[RC-1136]:: Adding channel dimension to MI dataset (#133)

[job-creator] prepare release release-2.12.0

[job-creator] prepare for next development iteration

Cast Timestamp filters properly into Timestamp objects

Summary:
Prior to this change, a UNIXTIME_MICROS field was passed in as a
string. And this mostly worked, but it caused
`FilterProjectTransposeRule` to blow up because the type in the query
`String` didn't match the type in the `Schema` which is
`Timestamp`. This resolves that.

[job-creator] prepare release release-2.12.1

[job-creator] prepare for next development iteration

Changes to server to allow unixtime_micros for reverse sorted fields

- Allow UNIXTIME_MICROS as datatype for datetime field, in case buildCondition logic asks for dataType
- Change QueryContext constructor to receive accountSidField and dateTimeField, thus allowing helper methods in the class to skip them as method args
- Change tests to reflect the same

Add in new resource and event_types to audit event mock-data

Table config in yaml template