Add a few more probes into Execute Log
Summary:
Still working on figuring out where all the time is being taken on a
given RPC queries to Kangaroo Jack. It doesn't appear to be sql
execution time as we can run the exact same query and get our response
far quicker.

The metric of interested in my testing is
```
jdbcConnectionAcquiredDurationCurrent
```
This metric is the time between when the Task to execute the SQL is
created and the time it has finally acquired a JDBC connection.

In my focused testing I determined that the JDBC pool returns a
connection very quickly. I used to measure this metric within the
`Runnable` task. This indicated ~1 millisecond of latency to acquire a
connection which is not where the time is going. So I moved the start
of the time up to the Task creation and we can now see it takes a long
time for a thread to pick up the task and execute it.