Transform an OUTER JOIN into an INNER for filters
Summary:
When we query for Transactions and filter by billable item category,
we see that the filter isn't pushed down into the scan of the BI
Category table. This is documented here:
https://wiki.hq.twilio.com/display/DATAPLAT/SLO+Report+for+2019-10-27+through+2019-11-04#SLOReportfor2019-10-27through2019-11-04-UsageReportTransactionsinvestigation

When we change that to an INNER JOIN it performs considerably faster
and so this rule does exactly that.

RC-1165

This also matches a changeset against Calcite:
https://code.hq.twilio.com/gist/sreynolds/054166858de86813d3ecb1e502626d03