Rollback OuterToInner join rule
Summary:
This rule is causing the planner to spend a lot of time
planning. Removing the DUMB filter join rule resulted in the KuduSort
not being applied on the left side of the join degrading performance
of our queries. This resolves it all for now by removing the
OuterToInner until we can piece together why SortJoinTransposeRule
only applies OUTER joins and not to Inner joins

RC-1203