Push A Sort past the Aggregation when sorted by primary keys
Summary:
This introduces a new variable `groupByLimited` to
`SortableEnumerable`. This boolean indicates whether or not the
`groupBy` implementations -- currently only implemented one of them --
is responsible for limiting the results and exiting early out of kudu
rpcs.

This also adds Fact, Cube and Dimension table row estimations this was
done so the cost of sorted aggregation plan is less then (not equal
too!) the plan generated without the sort. This happens because our
rule doesn't remove the original sort, it preserves it unlike the
other sorting rules

RC-1186