Only apply `AggregationTranposeRule` if goupby matches primary key
Summary:
This rule is exploiting the order in which the rows come out and
counting the unique group keys that come back from these rows. Because
of this, the group keys need to match the sort order otherwise we can
get unique group keys early and causing the execute to examine far
less rows that are actually matching the query.