Handle the last group by key properly
Summary:
Prior to this change, the last group by key that came from the source
enumerator would cause a duplicate group by key to be inserted in the
queue and have it's accumulator updated with it's value. It would use
`lastKey` which is the key that _was last submitted to the
queue_. This only affected queries that were not able to meet the
`fetch` limit. So queries that `LIMIT 1000000` and there is only 10
matching group keys.

RC-1209