extract the byte fields and ensure they come up in right order
Summary:
Prior to this change, the result wasn't selecting out the byte
field. The jdbc result set happily pretended that the 2 field was a
`byte` even though it was a long.

By adding the byte field in the select clause we are able to assert
that the result is properly sorted (6 should come before 4).