Simplify if statements to make logic clear
Summary:
The central concept, an accumulator is non-null (and has at least one
value applied to it) if and only if we have skipped enough records and
the last record we had hasn't already been published to the queue.

We can assert this because:
1.) accumulator is initialized in only one spot, after the skip has
occurred
2. After initialization, it is immediately used on the object.
3. accumulator is `null` at the beginning and after a new key
has been pulled from the source enumerator.