Move `matches` implementation into a predicate for ConvertorOperand
Summary:
The `matches` implementation was causing an assertation error to be
thrown by the `VolcanoPlanner`. The assertation was making sure the
rule actually matches the current call. This assertation is expected
to pass as it was already ready checked *PRIOR* to inserting  the
match onto the queue.

TODO: we need to look hard into *why* we get a KuduProjectRel as input
into another KuduProjectRel