Cast non-BigDecimal input literals for DECIMAL columns
Summary:
With the filter work, we will allow users to filter the chart and
reports on DECIMAL types. When a user inputs "7", Jackson Deserializer
returns that as a Integer. Prior to this commit that would cause the
`toPredicate` method of `CalciteKuduPredicate` to error out. This
updates that logic to handle Integer and Long types for Kudu Columns
that are type DECIMAL.

RC-1221