Return the ISO 4127 currency code in metadata response
Summary:
After doing all the relbuilder work, this regression was introduced.
The metadata no longer contained the ISO currency code. The cause
of the bug, was leveraging the `firstRow` of the response. Each row
 -- by design -- doesn't contain the `units` integer therefore,
 the `units` field could not be found in the
 `Map<String, Object> firstRow` !.

```java
.filter(columnName -> CurrencyUnitsMap.isValidUnits((Integer)firstRow.get(columnName)))
```
would always fail.

This change pulls out the `units` field from the first row from
`Calcite`, and passes it into the `setMetadata` method.

RC-1416