Re enable smart FilterJoin Rule
Summary:
This change adds to our Transpose rule logic to push down a sort
above a join even if that is an INNER join. Before we were relying
on `SortJoinTransposeRule` which only works on LEFT or RIGHT joins,
but could and should work on INNER joins as well. This creates a
version of that rule that applies to INNER joins.

This should improve the three way join query generated from Audit
Events when joining Product and Actor Dimensions

RC-1420