Use `AtomicBoolean` instead `CompletableFuture#cancel(boolean)`
Summary:
Calling `cancel` wasn't working out very well. It is because
`cancel` is called on the `leaf` of the completion chain not on the
root. This made all kinds of headaches so instead we using an
`AtomicBoolean` which is created when the HTTP request is
received. This boolean is passed through the framework into the
`KuduEnumerable` and checked. When it get's set we get an empty
`ExecuteQueryResponseContainer`. The code checks the
`QueryPlaner#isCanceled()` method and if it is set, returns a 504
back to the caller.