When using `DEBUG` log level on Volcano Planner queries would fail
Summary:
When using `DEBUG` level on Volcano Planner, would cause a litmus
test to fail. I believe that is because `SortTransposeRule` from
Calcite proper is producing the same plans as
`KuduSortTransposeRule`. There was a lot of redundancy after
enabling `smart` `FilterJoinRule`. This commit removes that
redudancy:
1. The `FilterJoinRule` will take a plan with Sort _over_ Filter
   _over_ join and make it a Sort _over_ Join _over_ Filter so we
   don't have to do that work anymore
2. Because the `FilterJoinRule` will push the Filter into the join
   it will leave the Sort above, therefore matching the original
   `SortTransposeRule` so our logic is now just redundant.

Therefore, we are able to remove these two rules and write a single
rule specifically for INNER joins. This makes the code considerably
more readable.