Place projection logic and function expressions into ScannerCallback
Summary:
This code simplifies the `KuduProjectRule` by creating a `Function1`
that takes as input the `RowResult` from KuduRPC and constructs the
`Object[]` that represents the row data. This `Function1` can also
handle all the function calls used in our projections (flooring on
DAY, CONCAT the fields together with `||` and handling `CAST` to not
null).

This lays the ground work for adding a `Predicate1` into the
`ScannerCallback` that will check the `RowResult` and make sure it
matches predicates that can't be pushed down into the Kudu RPC -- e.g.
`NOT EQUALS`, `LIKE`. By pushing these into the `ScannerCallback` we
can we can ensure the `KuduProjection` gets pushed into the plan as
well.

RC-1602