Add Support for all RexNodes in KuduFilterRel
Summary:
This change changes how we handle `RexNode`s in a
`Filter#getCondition`. Prior to this change `KuduFilterRel` only
handled `RexNode` that could be pushed directly into the KuduRpc. When
a `RexNode` couldn't be pushed down, `KuduFilterRule` placed a
`LogicalFilter` on *top* of the `KuduFilterRel`. This `LogicalFilter`
would be transformed into a `EnumerableCalc` which would use
`RexToLixTranslator#translateCondition` to do the filter in memory.

This change now has `KuduFilterRel` accepting all `RexNode`s and
provides all the information `KuduToEnumberableRel` needs to construct
a `RexProgram` with the provided `RexNode` condition and it is able to
hand this over the `RexToLixTranslator#translateCondition` to create a
`Predicate1<Object>` that can pushed all the way down into the
`ScannerCallback`. This `Predicate1` is then used to skip records from
the RPC that don't match it.

```java
final RowResult row = nextBatch.next();
if (filterFunction != null && !filterFunction.apply(row)) {
  continue;
}
```

RC-1602