Enable Calcite to push the Sort through the join
Summary:
Enumerable joins are able to sort with the following rules:
```java
// The current implementation can preserve the sort order of the left input if one of the
    // following conditions hold:
    // (i) join type is INNER or LEFT;
    // (ii) RelCollation always orders nulls last.
```
So this change removes the `JoinCommuteRule` which was biasing the
smaller table into the left slot preventing, this was preferable to
the optimizer because the smaller table hand large advantage. After
removing the `JoinCommuteRule` the `ENUMERABLE_MERGE_JOIN_RULE` was
the chosen join algorithm and this prevent the sort from being pushed
because it needed the table sorted on different fields. Removing these
too should improve our query performance on Fact table scans and
Pagination queries

RC-1463
RC-1611