Adds the IN equality type for filters
Summary:
This adds support for `IN` Equality type and add support for pushing
this down into Kudu. This allows the RPC to use this instead of a lot
of ORs, resulting in a lower amount of scanners and less memory
pressure. This is useful for messaging and for our own incoming
features.

RC-1629