Rework `IN` implementation to reflect Calcite's direction
Summary:
The previous change that added `IN` support was a hack. It used
`ARRAY` to construct a weird filter. This filter is not actually
supported by `EnumerableCalc` and when Calcite is unable to push
this filter down into the Kudu Scan it just breaks.

This filter -- added by the `QueryPlanner` also produced invalid
SQL, SQL that could not be taken from the `System.QueryLog` table
and executed in the `kudu-sql` shell.

This hack caused a query for Messaging Insights to fail with the
following error:
```
cannot translate call IN($t9, $t15)
    at org.apache.calcite.adapter.enumerable.RexToLixTranslator.translateCall(RexToLixTranslator.java:763)
```

From Calcite's mailing list:
```
We have had several discussions over the years about how to represent IN-lists (e.g. “x IN (1, 3, 5)”) in RexNode-land.

I have generally taken the position that we should expand to ORs (e.g. “x = 1 OR x = 3 OR x = 5”) but a few months ago accepted that we should allow IN in RexCall.

I have given this some further thought, as part of a couple of RexSimplify bugs [1] [2] and I now
think we should replace IN with something more powerful, namely sargs. A sarg (or search argument
[3]) is an ordered set of intervals, and can be represented as a Guava ImmutableRangeSet, such as
"[[0‥1), (1‥2], [3‥3], (5‥+∞)]". It can represent an IN-list of constants, but also ranges.
```

https://mail-archives.apache.org/mod_mbox/calcite-dev/202008.mbox/%3CDF328A3B-769E-4293-8924-3F4D6E07B2E5%40gmail.com%3E

Given SARGs is not implemented, this uses the position of expanding
it to ORs.

It achieves the in list optimization  by iterating over all the
`Scans` -- `Scans` are represented as `List<CalciteKuduPredicate>`
-- to find columns that each scan has an `EQUAL` condition on.
For these columns, the code collects all the unique `EQUAL` values
and constructs an `InListPredicate`.

RC-1709