Fix comparison operator being created for column 0 each time
Summary:
Prior to this change, whenever `InListOptimization` found a column
that could be optimized *but* only contained a single value it
would hardcode a column index of 0. This produced a
`IllegalArgumentException` when creating a `KuduPredicate`:

```
java.lang.IllegalArgumentException: account_sid's type isn't [Type: int8, Type: int16, Type: int32,
Type: int64, Type: unixtime_micros], it's string
```

This is because we were, in the specific case, trying to use an
integer of `1` as `provide_feedback` and that value was present in
both scans *but* had the same value. Because `0` was hard coded the
new `ComparisonPredicate` referenced the first column in the table
`account_sid` instead of the `provide_feedback` column.

RC-1709