InList Optimization shouldn't change an AND to an OR
Summary:
InList optimization introduced a regression in how we handled one
query for Messaging Insights -- a query that will always return 0:
```sql
WHERE STATUS = 'DELIVERED' AND
  (STATUS = 'UNDELIVERED' or STATUS = 'ERROR')
```

The previous code would turn this into a single InList of:
```sql
WHERE STATUS IN ('DELIVERED', 'UNDELIVERED', 'ERROR')
```
as it wasn't written to handle a scan that contains the an EQUAL
filter on the same field in the scan. This is obviously a query
that will return an empty result set.

This change brings back the behavior that was present before
```
EnumerableAggregate(group=[{0}], COUNT_RECORDS=[$SUM0($1)])
  KuduToEnumerableRel
    KuduProjectRel(DATE_CREATED=[$1], COUNT_RECORDS=[$10])
      KuduFilterRel(ScanToken 1=[account_sid EQUAL ACc781c131b82e694c702622de4f9eb114, date_created GREATER_EQUAL 1597795200000000, date_created LESS 1598399999000000, status EQUAL DELIVERED, status EQUAL FAILED], ScanToken 2=[account_sid EQUAL ACc781c131b82e694c702622de4f9eb114, date_created GREATER_EQUAL 1597795200000000, date_created LESS 1598399999000000, status EQUAL DELIVERED, status EQUAL UNDELIVERED])
        KuduQuery(table=[[kudu, OutboundMessages-SubAccount-Aggregation]])
 ```

This has two scans that query:
1. `DELIVERED` and `FAILED`
2. `DELIVERED` and `UNDELIVERED`

RC-1709