Leverage `java.lang.Number` for InList predicate casting
Summary:
Prior to this commit, when handling `INT8`, `InListPredicate` assumed
the value coming from Calcite was an `Byte` but it is actually an
`Integer`. This caused a `ClassCastException`.

This change casts the object from the `RexLiteral` into a `Number`
and then calls `byteValue()` to translate it properly.

This commit also adds tests for all the Schema types to protect
against this happening again.

RC-1796