Travis setup to deploy artifacts
Summary:
Staging is resolved and we are now ready to execute the release

This adds a settings.xml which sets username and password from
secure environment variables from travis pipeline

Artifact is signed using simplified maven plugin that grabs gpg
password from secure travis environment variable.

The private key is encypted with a Private key known only to Travis
and stored in `travis/secrets.tar.enc`. Travis hands the build the
encryption key at build time and it never lives on a laptop.