Fix generated javadoc
Summary:
The javadoc this repository was generated created several errors and
quite a large number of warnings. To resolve this, this change fixes
several of the errors in javadoc and then ignores generated parser
code.

Ideally, the parser code generation is fixed to generate valid javadoc

All future changes to this repository will fail to build for javadoc
warnings or errors.