Simplify sort rules to remove complex logic
Summary:
These rules feel very tortured. This simplifies the operand a lot and
preps it for moving to a Converter rule.

It simplifies the operands and leverages the
`RelMetadataQuery#getAllPredicates` to surface up the filter
conditions. This should allow it to be more generic and applicable.

TODO:
Need to come back around to the sorted aggregation rule and adjust it
to use the new sort aggregation in kudu enumerable rules