Apply sort when strictly filtering on non-pk column
Summary:
When a query is sorting on a non-primary key column *AND* that query
is filtering to an exact value for that column, the sort rule can be
applied.

These queries are a nonsensical, a sort is applied even though it has
a single value. Ideally,
[ReduceExpressionsRule](https://github.com/apache/calcite/blob/404f968f6079360621896e6ee20ec12488a40bc2/core/src/main/java/org/apache/calcite/rel/rules/ReduceExpressionsRule.java#L696)
would remove this sort as it doesn't make any sense. That is a larger
work to Calcite