Revert Sort Simplification
Summary:
As noted in the last comment in this PR:
https://github.com/twilio/calcite-kudu/pull/2#issuecomment-754976144

The rule is failing to properly handle a Sort Over a Projection. When
Sort is over a Projection, the `sortField.getFieldIndex()` reflects
the field's position in the Projection and the code assumed it
reflected the field's position in the Kudu table.

This reverts commit bd9f9055a074afae0d940deaae8edaaad7c375b6, reversing
changes made to 0b44ba05eceaa238b94fcde6132c82244ed7ee7d.