Build a Caching Nested Join implementation
Summary:
Today, nested joins can be very slow creating the same Scanners for
the right table. It is common, for instance, to fetch the same right
rows after each batch from the left.

This attempts to cache some results from the Right scan that match the
left side.

The `KuduRowCache` maintains a `LinkedList` of cached rows and it uses
a `EqualityComparer` to reduce duplicates and to power comparisons.