Caching Nested Join implementation
Summary:
This implements a cache between the resulting Predicates when
executing a nested loop join.

Problem:
In a Nested Loop Join, the query is sending the same RPC requests over
and over again and fetching the same results. This makes the query
perform notably slower because of all the network traffic.

Solution:
This creates a fixed size `Map<Predicates, CachedResults>` that is
local to the current nested loop join context. Each `CachedResult`
maps to a set of predicates specific to the row on the left and
therefore shouldn't return thousands or millions of results -- it is
assumed the cached results will be small but not necessarily uniform
in size.