Make the cache use access order
Summary:
Access order makes `get` operations intense as the linked list is
shuffled around as the key is moved to the front of the list. This
does give an LRU behavior for when all the scanners cannot be held in
the cache. The evicted key is the one that was accessed the earliest
in time.