Use the max Cube granularity among the cubes
In case of mixed time granularity of cubes and using min amongst them for deciding on the number of threads, the time ranges are unique for the smallest time granularity cube but not for any cube having granularity greater than that.
Hence, rows get overwritten(upserted) in CubeMutationState from different threads for any cube bigger than the smallest granularity.
In order to keep uniqueness in thread splitting of rows to write, we need the max granularity rolled up time of a cube

Code formatting change