Simplify KuduSort and push the limit down more often
Summary:
While reviewing #19, we decided the logic around *when* a limit is
pushed into the scans, it can be pushed in even when there isn't a
sort. This does produce weird results as it is a race to which
scanner completes first -- but that expected when there isn't a sort.

This also fixes a bug when `filterFunction` is something else for
instance, `NOT EQUALS` or `LIKE`. In those instances, the limit should
not be pushed down