Reduce cost on every `CLOSE` message when logger is > DEBUG level
Why:
Looking a bit at query performance for a `null set` query I noticed
this code was calling out to the `AtomicBoolean#get()`. This uses a
votiale:
https://github.com/openjdk/jdk/blob/master/src/java.base/share/classes/java/lang/invoke/VarHandle.java#L579-L606

Which can be "expensive" to reference.

How:
If the logger is configured at `DEBUG` level, then the logger will
call `AtomicBoolean#toString()` which in turn references the
`volatile` to render the String. So this doesn't change behavior if
the logger is `DEBUG` level but does prevent the `volatile` from being
dereferenced if logger is configured for `INFO` or higher.