`KuduFilterRel` calculates cost based on partition pruning
Why:
To improve the costing metrics of different plans, the `KuduFilterRel`
should cost itself based on the number of partitions out of the total
number of partitions. This calculation has to be cheap and hopefully
done once

How:
The `KuduFilterRule` now pulls out the `KuduClient` and uses it to
make a `ScanTokenBuilder` for the table without predicates and makes a
`ScanTokenBuilder` for the table with the predicates. These two sizes
allow the `KuduFilterRel` to calculate the percentage of partitions
removed. This is used to reduce the cost of the plan based on that
metric.