log4j1.x removal
Why:
Log4j1.x has been end of life for a long while.

How:
It was added as a test dependency so we could control the log output
during builds. I lazily added log4j that ships with slf4j -- which is
log4j1.x. After moving to log4j 2.15.0 implementation, we can see that
https://javalibs.com/artifact/com.google.uzaygezen/uzaygezen-core
brings in and old log4j library -- and it brings in an old apache
commons and guava. By specifying log4j as an `exclusion` it will not
be included into the dependencies.