Set commons-io to 2.11.0 to address CVE-2021-29425
Why:
`calcite-core` pulls in 2.4 which was released in 2012. Versions less
than 2.7 allow strings to traverse into the parent directories.

https://security.snyk.io/vuln/SNYK-JAVA-COMMONSIO-1277109

How:
Adds a comment around the `calcite.version` property to help the
engineer check the calcite bug to see if it is addressed.

It adds `commons-io` dependency into the `dependencyManagement`
section so that it is used whenever `commons-io` is pulled into a
module.