Fix commons-dpcp2 vuln and clean up other vuln fixes
Why:
commons-dpcp2 used to leak password information over jmx. This causes
Synk to flag it as a problem.

Looking at the other fixes, after bumping to calcite 1.31.0, we can
remove the json smart, httpcomponents and jackson changes

How:
Adds a commons-dpcp2 to dependency management with the latest
version. This ensure calcite doesn't pull down the vulnerable
one. Added a links for the resolution and added a note about above
calcite version to remind us to clean this mess up once resolved in
calcite.