Re-enable Unique Semi Join rule
Why:
After working on CALCITE-5391, I came to the conclusion the rule is
not actually broken -- not strictly -- and Twilio's own application is
actually not using Calcite correctly. As discovered by Drill and
Twilio, Calcite's planner will accept a new `RelNode` from a rule that
matches the shape of the original *but doesn't have the same
names*. This allows for different optimizations within the planner.

To use Calcite correctly, the applications need to keep the original
validated type from before the physical planning. This can be seen in
how the JDBC uses RelRoot.

How:
Removes the removed rule